Update BackupScript.ps1
Bugfixes
---------
Fixed a bug with the missing attachment when $SendMail is true.
Fixed a bug where the stream could not be read. Now the logging function uses the StreamWriter class.
Fixed a bug where it would not copy files in the root folder. The mechanism with getting the folders first and recursing files for those after was not performant on my side.
Fixed a bug where it would display an error if the $Backup folder does not contain any files. It still copies backups the empty folder.
Fixed a bug where it would should the incorrect file counter in Write-Progress and log.

Improvements
----------------
Improved performance by adding a stopwatch to the Write-Progress line. So it only updates the progress after at least 500 milliseconds. If the backup is done it closes the Write-Progress bar.
Changed some functions names to match Get-Verb and Where-Object aliases.
Added the compression switch to 7zip.